% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvr_correct_url.R
\name{cvr_correct_url}
\alias{cvr_correct_url}
\title{Corrects url encoding for Danish letters.}
\usage{
cvr_correct_url(url)
}
\arguments{
\item{url}{A build url.}
}
\description{
Corrects url encoding for Danish letters.
}

