% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvr_get.R
\name{cvr_get}
\alias{cvr_get}
\title{This function makes a request to the CVR API.}
\usage{
cvr_get(search, country, search_in = NULL)
}
\arguments{
\item{search}{The seach will be made in the CVR-number, p-number or company name.}

\item{country}{Specify country to either 'dk' for Denmark or 'no' for Norway.}

\item{search_in}{Not required. Character value 'vat', 'name', 'produ' or 'phone'.}
}
\description{
This function makes a request to the CVR API.
}

